Inline _refresh_ini_tweaks -> refresh:
_diff_dir put to use - see with whitespace off
- _refresh_ini_tweaks: I replaced naked factory call with new_info
which necessitated passing booting to refresh (aligning refresh
signatures used presently). I used in in RefreshData but not in
ShowPanel - again priority is RefreshUI -> ShowPanel will follow (see
- files_to_delete: if the info was corrupted (deleting via BAIN?) we
still need the factory call to use the abs_path operations. This fixes
a bug actually as the key error would be thrown in the iteration - the
use case is uninstalling a corrupted mod/ini/bsa (?) - untested but more
probable it works now. This necessitated returning a dict from
filter_essential. The rest of the uses are through the UI so we won't
have any Nones there. This was ripped from "Revisit delete" commit - see
there for using the dict and more such refactoring.

No need to filter new_or_present as delete_refresh does that?

            for k in list(self): # todo test - should be redundant
                if k not in new_or_present: del self[k]

Align refresh:

Fixes what seems to be a bug me - we should not notify BAIN for
resetting of default tweaks - the deletion of the copy should have been
communicated in delete_refresh. Would blow with an AttributeError on
abs path sans doute.