AFileInfo - decouple Screens:
We do not want to add a table to ScreenInfos, image handling should stay
at a minimum it's trouble enough already :P So this decouples screens
from TabledFileInfos (table itself must go but that's 314+).

Notes on copy info:

- return value was never used (since ...)
- Only place set_mtime is used is _CopyToLink.Execute. This permitted
to move the setmtime call inside modInfos override and move ScreenInfo
higher up - has no masters and no backups (which is a hack sprinkled
here and there that we need to revisit and certainly not leak to the
rest of the Infos).

WIP #336.

Infernio:

Fix duplicate on default tweaks

This used to work (I know because I made it work when I improved the INI
Edits tab's usability), so this is a regression. Not sure if this is the
right to do it though, the rename APIs are still arcane to me :(

Ut: the fix for default tweaks was moving validate_name in ListInfo -
the previous behavior when copying a default tweak to a file was to
create a file with the same name as the default tweak and this is
reflected in IniInfos refresh - we should not add the 'Copy' suffix by
default - if we don't add 'Copy' currently breaks with an attr error on
abs_path.

Drop GPath(fullpath):

Unfortunately this:

@@ -1814,5 +1814,4 @@ class AFileInfo(AFile, ListInfo):
     def __init__(self, fullpath, load_cache=False, **kwargs):
-        g_path = GPath(fullpath)
-        ListInfo.__init__(self, g_path.stail) # ghost must be lopped off
-        super().__init__(g_path, load_cache, **kwargs)
+        super().__init__(fullpath, load_cache, **kwargs)
+        ListInfo.__init__(self, self._file_key.stail) #ghost must be lopped off

won't work as the fn_key attribute is used in ModInfo init for instance:

  File "D:\...\Mopy\bash\brec\mod_io.py", line 280, in from_info
    return cls(mod_info.fn_key, mod_info.abs_path.open('rb'))
               ^^^^^^^^^^^^^^^
AttributeError: 'ModInfo' object has no attribute 'fn_key'

I anyway removed the GPath from there - let it blow if we pass a str
(note the fullpath.root in ModInfo.init), plus we GPath in parent (...)

Mopy/bash/ini_files.py: IniFileInfo(AIniInfo, AFileInfo)

MI here is real tricky but I want to have this inherit AFileInfo like the
rest - we could raise on get_rename_paths but that's just extra lines.