Work around icon size crash
Still no fix in sight for this - it seems to have worked for a while,
but now it's broken again. Work around this once and for all by
requesting a restart instead.

Ut:

It was easier to fix than to explain all the subtle ways this would
go wrong (if user unhides a button that was hidden in previous session
and button is created with updated icon size) - fixes status bar being
redrawed with updated (wrong) icon size if user chooses to not restart
after changing icon size. The assumption that icon sizes could be hot
switched was hardwired in the code so I had to hardwire the opposite.
Note BashStatusBar was in scope at least. This was a showstopper as
(not) discussed - thanks :)

This triggered some (final) refactoring of the status bar - turns out
we were performing a lot more refreshes than needed and in particular
we shouldn't bind the EVT_SIZE - maybe this was causing the crash?

Closes #334 for now - only a workaround, we may return to that issue.