Fixup for displaying the CS button even when the CS is not present:
There was an easy fix (duplicating the logic of app_button_factory in
TESCSButton init) and a correct one - add more per case logic where
there is some already and keep call sites of find_launcher at 1.
Patience till we deprecate the app_key (None in this case).

Corrections to the other soon-to-be-deprecated readme from
https://discord.com/channels/537706885965676554/537710082755133460/1179863067350270073

Correctly passing the cli args to xEdit:

We were reading it as one argument, see
https://discord.com/channels/537706885965676554/537708965942001685/1181998141583655022