Cleanup status bar text apis:
We had:

- BashFrame.set_status_count used once in
- SashUIListPanel.SetStatusCount used twice - is the if needed in
  ShowPanel too? - Inf: nope, since ShowPanel switches to the tab in
  question
- BashStatusBar.set_sb_text the real one (wraps _native), but why did it
  call self.refresh_status_bar()? -> turns out we needed this on show_panel
- BashFrame.set_status_info, the API

This cures the flickering of the status bar buttons while not ellipsizing
the third sb field (item counts). The flickering was due to SetStatusWidths
that was only needed when the panel was shown. Not sure if we ever bypass
the ShowPanel call - onSize was not needed here.

Mopy/bash/balt.py:
Removing this call entirely does not seem to cause problems on Windows?

Mopy/bash/basher/app_buttons.py: Guess this is the intended behavior?