Use an AFile for loadorder/plugins txts: TTT
This is a mini branch actually - contains also some refactoring of
get_*file, which were moved to base and _backup_*** overrides were absorbed
(for the IniGame based on the observation that the _handle_*** test is
performed in the get_*file). It is WIP in the sense that those mirror
methods should be atomic but more on that latter. We could also probably
set the ini files to the AFile attributes hmmm. Note error handling is
unified in __backup.

Needs testing especially: TTT

- _plugins_txt_modified now returns True for deleted plugins.txt
- same for load_order_changed for deleted load_order_txt
- swap uses FNF untested
- ...