request_cache_update:
Adds a few lines but replaces a pair of functions with one and removes
redundant _plugins_txt_modified. Simplifies client code as it splits
responsibilities - GameLo should know the interaction between lo and
active, the client just needs the args to pass to get_load_order.

Use the LordDiff machinery for cleaner code:

- _cached_lord.active != saved.active - yep active order does not matter
cause this would mean load order changed so `or` would be
short-circuited. Still let's be expressive here and call act_changed -
refresh_lo is actually the stitching holding load_order.py together, so
keep it explicit (cf the docs in mods_links). Note we pass the fixed
saved load order at this point - and _cached_lord is updated in
_update_cache. So deletions should be correctly not included in
ldiff_saved - for Timestamp games as the lo is always recalculated
based on modInfos, for the rest we need to correctly set cached_values
to None. So the complexity bubbles up to refreshLoadOrder - we should
be passing a RefrData-like in there to make it more traceable. Additions
should trigger a deprint if the _cached_lord places them anywhere but at
the end of the list - can happen for Timestamp games - so saved (fixed)
places them at the end while get_load_order (_update_cache) would put
them in mtime order. I don't mind that deprint but when I do I would
need a more granular input - laterz.