Move static load order overrides to PatchGame:
Should probably be one level up but that's enough for now - takes down
one nasty use of fsName - could do this with more of the factories but
*not yet* - this was the lowest hanging fruit. Other than that _games_lo
lost one third of its size - it would keep growing with new games which
is a no-no - it's a library.

We could consider moving max_epmls to a GameInfo class variable and add
it as an init parameter to LoGame - then the load_order thin wrappers
would become redundant. LoGame and GameInfo just met.