exePath -> app_path:
We need to pass kwargs from a dict - and these stay for ever in
settings - time for some PEP8. I chose launcher_path - still not final.

Those finally: self.... = old was a clear sign that we were doing
something wrong - turns out we needed `exePath` in special occasions
and calculating it was completely negligible performance in those
cases. The solution is this app_path/app_cli property (the latter was
already there as exeArgs) - this took down a few sb_click overrides and
one init override.

I used exepath in sb_click, only real use of app_path apart from the
Tes4Edit weird one in links_init. Note we want _app_path in
_app_version - or so it seems. Note also that if app_path != _app_path,
the latter is present and the former not, we get _display_launcher True
and boom in sb_click - that's why app_path comes with a don't-override
warning. Some more thought is needed but meanwhile app_buttons is
officially no longer ugly - under #570.