Fix fatal errors and warnings in translations
This allows msgmerge to run successfully on all of them. Tried to
preserve as much of the broken translations as possible.

What is up with that Tree[d] thing? It seems to have been there for
ages. I removed the backslashes to ensure the translations don't break
again from this.