Return the diff from _cached_lord in case we revert to saved load order:
I initially returned saved.lo_diff(fixed) in that case but this led to
discrepancy between rdata and ldiff, something to be avoided. For
instance when booting rdata.to_add is all the plugins in the data dir,
but ldiff.added may be a {new plugin} whose mtime would not put it at
the end of the load order, where it is put when fixing the saved load
order - so ldiff_saved would be returned which would only contain the
new plugin - similarly, ldiff.missing could differ from rdata.to_del.

Debug messages could use some work but once LDiff is frozen (and merged
with FixLo)

Drop setGhost from _activate (yey):

_modinfos_cache_wrapper should do that if mods were unghosted since it
does not rely on autoGhost anymore