Refactor refreshTracked - fix refresh glitch:
Clarify a bit - trying to figure out if we can do the refreshTracked
call before the irefresh one so we only call refreshStatus inside
irefresh - see basher here. This is also related to an elusive refresh
glitch - namely, uninstalling then installing an archive with one esp
sometimes on tabbing out/ back in to Bash would start showing the esp
as missing (so package turns red - autoghost is on not sure if related)
- goes away with anneal - well sometimes. Seems it is due to the
refreshTracked call being delayed and unconditionally popping
del_paths. To actually *not* pop it in ShowPanel we must call
refreshTracked in _removeFiles -> done. Note ghosts are added here.
Note also:

@@ -2497,2 +2497,4 @@ def refreshTracked(self):
                     altered[apath] = tracked.fsize, tracked.ftime
+                    # if we uninstalled then reinstalled without leaving Bash
+                    del_paths.discard(apath)
             except OSError: # untrack - runs on first run !!

does not solve the bug as the esp may (won't) be in _miscTrackedFiles
but won't harm.

Inline refreshInstallersStatus

We always get a set in install AFAICT - I removed the set call.

I am not a big fun of local functions (_path_key) - but in a do once
operation is okayish - in this case I wanted to cache the stats, so
it's justified. However the real solution is the data tab (#50) - also
an AFileCrc could come handy.

Unify irefresh calls in ShowPanel:

We might be able now to add the progress back in converters - but
should it be part of irefresh anyways (edit: nope infinite refresh
still!)? Note that now the `elif self.frameActivated` branch also sets
self._data_dir_scanned to True (same with _user_cancelled on
CancellError) - should be ok.

Under #219, #353.