_ask_info:
Reduce data_store occurrences - data_store is a very important attribute
and the only thing this repetitive logic was conveying was that the
labels/sort keys depend on the info. The downside is getattr, but we
can live with that for the readability - which makes more apparent the
differences of Installer vs FileInfo and also some common patterns - so
this is WIP but will do for now (we should add a size_str and time_str
method to AFIleInfo - from future, #336). Also note in SaveList the
static methods - I don't see why a save info should have no header -
needs some testing as I bit the bullet and drop the if header checks.

Remove cached_lo_index_or_max:

Only use remaining -> get_ordered - the latter should have a variant
that accepts missing mods (when sorting mod/save masters that might be
missing for instance) - in the rest of the cases all present,
non-corrupted mods should have a unique _cached_lord.mod_lo_index so in
particular the ones displayed in modList hence

@@ -970,4 +970,4 @@ class ModList(_ModsUIList):
         'Installer' : _ask_info('get_table_prop', ('installer', '')),
-        'Load Order': lambda self, a: load_order.cached_lo_index_or_max(a),
+        'Load Order': lambda self, a: load_order.cached_lo_index(a),
         'Indices'   : lambda self, a: self.data_store.real_indices[a][0],
         'Modified'  : _ask_info('ftime'),

hexIndexString -> cached_active_index:

That's not modInfos that's directly load_order. The only use case was to
display that index actually - modifies default=default_index to have a
double 0 - it's okay

f'{0:X}'
'0'
f'{0:02X}'
'00'
f'{15:02X}'
'0F'
f'{15:2X}'
' F'
f'{15:0X}'
'F'
f'{15:X}'
'F'