ci: Automatically rebase and push weblate-in
Should hopefully work - if not, we'll have to do merges here and then
manually clean them up every once in a while when we merge changes form
weblate-out. Not the end of the world and could probably be automated
with a script if we did have to go that way, but let's hope this works.