l10n: Update translations for ellipsis too
Otherwise we'll get a ton of failing checks for those as well.

That's also why I have to keep the two commits at the start of this
branch unaltered - that way, weblate should skip over them when
rebasing.

In its own commit because it's simple regex replacement and github loves
choking on the l10n changes.