Standardize unopenable item filtering
We were doing it ad hoc in a link class. Just like essential filtering,
this should be done at the data store level. That way we can use it in
every place that wants to open files. Triggered by me getting a
traceback when I randomly pressed Enter while I had a marker selected on
the Installers tab.

Takes two tab-specific link classes with it, which is very nice.