Add support for displaying GIFs in gui
There was a report that an FOMOD containing a GIF failed with a wx
error:

    wx._core.wxAssertionError: C++ assertion ""image.IsOk()"" failed at ....\src\msw\bitmap.cpp(872) in wxBitmap::CreateFromImage(): invalid image

This turned into a major undertaking. This still isn't anywhere close to
perfect (most notably, GIFs don't resize properly and will instead just
cut off), but it's far preferable to crashing WB or breaking the FOMOD,
which was the old behavior. It could have been a lot simpler if
wx.Animation.GetFrame and getFrameCount weren't *unimplemented* on GTK.
In that case, I could have just done the frame advancement myself and
integrated it properly into the existing implementation. Maybe an
approach to solving this is to pull in a third-party dependency for
reading GIFs that is independent of wx.

Architecturally, we also have a lot of improvements here. The various
caches moved into one central cache that also caches by CRC. This also
thinned ScreenInfo to the point where the class has no reason for
existing, so we could (read: should) probably drop it now.

Under # 190 and # 336 <--- RRR

Ut: Lo's comments on handle_resize: Thanks :)