Refactor rename_operation:
Again the key is return what's needed (once one figures what this is) -
finally got me rid of _try_rename. Still WIP (RUI in particular which
is the goal here) - we need to reduce try_rename overrides, but it's
good for now (note the Installers one is now aligned with the rest -
and fixed as the len of refreshes was not necessarily 2 as the
`return False, False, False` seemed to suggest). That's under #353.

Under #580, #336