Restore support for disc/retail versions of games
Some of the old games (Morrowind, Oblivion and Fallout 3 to be exact)
had disc/retail versions that were independent of Steam.
20dd955312a328000f217e9a84c9f20b8f8566c2 broke detection of those, since
previously they were included in the registry check we used for the
Steam version. Restore support, but don't duplicate paths for the GOG
and Steam versions (which set the same registry key as the disc
versions).