Allow relative detection to work with renamed Mopy
Right now it's checking the Mopy path to see if it's named Mopy, there's
no good reason for that as far as I'm aware. Plus we already have the
Mopy folder in bass.dirs (which is guaranteed already initialized since
init_dirs_mopy runs before we ever enter _main).

The rest are some minor renames and hunting down u''.