Allow editing case-insensitive BASH section
We accept a case-insensitive BASH in the description (see
getBashTagsDesc), but when changing the BASH via setBashTagsDesc, we
were still looking for a case-sensitive BASH. Look for a
case-insensitive one here (note that we will replace a non-uppercase
BASH with an all-uppercase BASH like this, that's fine and probably what
we want to do).

requirements.txt:
requirements-scripts.txt:
Update dependencies

Nothing major, mostly bugfixes.