Remove _files_to_delete:
Since we already override both _delete_operation and delete_refresh in
Installers. filter_essential is right where it should be, once. Yey, and
to celebrate I sprinkled a finally in there - I use it to mark areas of
the APIs that are final but I am not really a fan.

I had to add an fn_key attr to Corrupted otherwise:

  File "/home/infernio/Programming/wrye-bash/Mopy/bash/bosh/__init__.py", line 1651, in _diff_dir
    if inf.fn_key not in inodes}
       ^^^^^^^^^^
AttributeError: 'Corrupted' object has no attribute 'fn_key'

Corrupted is still new - it may inherit aunt ListInfo once the dust
settles.