Call refresh on data stores init:
This is three occurrences of precious refresh down and sets some very
preliminary machinery in prep of disabling the costly SaveInfos refresh.

Hacky fix for needing the modInfos singleton during modInfos refresh

This should need a completely different treatment - done later (this is
what sprang up this merge). Unfortunately modInfos is also used in
readHeader so we won't get rid of this soon.

Under #353