Refactor setOblivionVersion: RRR
This is one of the most arcane pieces of the FileInfos machinery.
Specification lost in Oblivion, file operations (# RRR 241 ) - reported
heisenbug stuff. Note size_esm_version[oldSize] was outside of the
KeyError try/except - some of this could be converted to EAFP?
Some of it moved to bush.game? What exactly needs be refreshed? Should
we disallow activating the modding esms or does not make sense/it's too
complex (_game_lo is actively refactored currently)? Black magic, I
reduced the repetition in the logic - this made me notice a bug in
error in 907bdccaea6e1474a433813a70cdbcddad28ae76 (i did not change the
parameters to infos in the error handler). I absorbed the helper so
this is as few lines as possible. As seen from questions above
setOblivionVersion badly needs a spec but still works (well see next
commit) - and hopefully now it's clearer what it does.
Mopy/bash/basher/mods_links.py: disable the menu if it won't do anything