Refactor setOblivionVersion:
Oblivion esm swapping is one of the most arcane pieces of the FileInfos
machinery. Specification lost in Oblivion, file operations (#241) -
reported heisenbug stuff. Note size_esm_version[oldSize] was outside of
the KeyError try/except - some of this could be converted to EAFP? Some
of it moved to bush.game? What exactly needs be refreshed? Should we
disallow activating the modding esms or does not make sense/it's too
complex (_game_lo is actively refactored currently)? Black magic, I
reduced the repetition in the logic - this made me notice a bug in
error handling probably introduced in
907bdccaea6e1474a433813a70cdbcddad28ae76 (I did not change the
parameters to infos in the error handler). The three next commits fix
more bugs and clean the logic up a bit but this needs handling of
store_dir to be finalized - see table merge. I absorbed the helper so
this is as few lines as possible.
Mopy/bash/basher/mods_links.py: disable the menu if it won't do anything