Drop BashApp RRR
I thought let's inline those two static methods as they're not as huge
as in the old days. Turns out 'Start application' was not accurate -
BashApp did nothing much anymore - the thin class so to speak. The boot
code is nowhere near to where it was 10 years ago, looks almost boring
now, used to be a perilous mess of wx and weird unpredictable stuff
with queer names like bosh. Hundreds of tracebacks. I can't say I luv
the splash thing but I won't remove it cause it is a screen. So I made
it more fail safe and less ugly, it's almost cute now that I don't need
to assign it. But what I like most is further untangling the main
modules responsibilities - remember initBosh? :P

Under # <--- RRR 600