Starfield: Add support for saves created by v1.9
The 1.9 update added another 4 unknown bytes to the masters list right
before the unknown short (i.e. right after the ESL masters). They seem
to always be zero. Who knows what Bethesda is doing there.

Also includes some more investigation - for example, I'm reasonably sure
that the second game version is the version of the game at which the
playthrough was started, since unlike the first game version, it does
not seem to increase when a new save for the same playthrough is created
by a new game version.