Add support for Skyrim VR ESLs
I have not implemented the whole backported ESL thing for regular
Skyrim. First off, I dislike supporting older game versions. But more to
the point, xEdit does not implement anything like it either, so I'm not
sure the whole description of it being HEDR-version-independent is
actually true.

Let's just operate off the assumption that SSE always supports 1.71 and
the expanded range and if that breaks anything for anyone it's a PEBKAC,
with the solution being either to upgrade to the latest game version
(recommended) or to install the backport plugin.

Also includes some minor fixes for to_os_path usages, Enderal SE updates
(latest Enderal SE on modpub supports HEDR 1.71 and expanded range, so
same situation as SSE now) and adding support for the limit fixer plugin
detection on Linux.

Introduce GameInfo.post_init

Turns out bass.dirs['mods'] isn't set by the time init() runs yet, which
makes perfect sense in hindsight. The exact point at which we post-init
is an open question - plus, in the future we'd like to use bosh.se_infos
to check for xSE Plugin existence etc., so that's another can of worms.

Ut: that's #600 - we 'd need to define boot stages (see
initialization.bash_dirs_initialized and co), but this needs the
dust to settle from #336 (includes initialization work). For now keep
post_init use count to 1 :)

Closes #673