ModInfo.str_bsas_sorted:
Drop _find_string_bsas in favour of a cache - should make loading files
faster (note removed performance comment) plus single responsibility
pattern - moving the handling of the key in the Bsa _str_heuristics per
game (SSE). Note the comment for SkyrimVR - comes from a future deeper
understanding of the situation and is small print for now - so I kept
previous behavior, as in using the heuristic load order. For most games
it coincides with the bsa load order as ini-loaded have negative lo,
but for SkyrimVR we might incorrectly ignore resource_override_key
bsas IIUC. I will fix that in "Use get_lo in getStringsPaths" commit.
For now note that ModInfo.str_bsas_sorted is set in
refresh > _refresh_mod_inis_and_strings, so will be None if we just
create an info outside the data folder or don't call refresh (or
new_info) - then if we wrap that in a ModFile and load -> boom. Probably
this should somehow be moved in readHeader - that's after #336 / #353

WIP ABsa.str_bsa_sort_key:

This is WIP as we will eventually move the handling in game subclasses,
and the str heuristics will only run for the MasterEsm(ModInfo) to be.