TTT get_bsa_lo -> single occurrence:
The previous situation looks like a bug: in the get_bsa_lo([self.fn_key])
call we would be adding to bsa_lo for each localized plugin the index:

bsa_lo[binf] = i (i would alwayss be zero here)

This meant that we would search for strings in some other plugins bsas.
I changed it so it copies self.__bsa_lo which at this point is just the
bsas that are loaded by inis - is this correct? TTT Also:

- why we use reverse - see next commit
- should we del available_bsas[binf.fn_key] in place or pass a copy - aka
is allowed to load strings from the same bsa from two files? (no I guess)

in short we'd need a couple more comments