Inline AsteriskGame._fetch_load_order:
Behavior change: we now resave if we detect any of rem_from_acti
whether we query for active or lo - from class perspective this makes
_fetch_load_order abstract which is correct - makes no sense for
AsteriskGame any more than _fetch_active_plugins - technically we
would need a different parent/mixin but that's an overkill.
_cached_or_fetch makes sense and will do - note we now respect the
contract of returning the cached_lo if this is passed in. Architecturally
this makes both _fetch_load_order and _cached_or_fetch single-use which is
as much as one can wish for - still test, as there might be some obscure
reason this should only be updated on fetching active plugins info.
_active_entries_to_remove did not do much and is inlined.