Drop the error checks:
Trust _voAvailabe, voCurrent are up to date

Check if move_to exists in our cache - this also catches the case
move_to be present and ghosted which the exists check would miss - we
have to be careful and check corrupted too (and trust those are up to
date too). API limits here.

Mopy/bash/game/__init__.py: let this blow with a KeyError - modInfos
when used must needs contain the master_esm.