Adieu, new_info:
refresh is the first class citizen here - since new_info really here only
updates the BsaInfos attributes which are not used in refresh itself we
might as well override and do all these updates in a loop. This makes
new_info completely redundant - on to delete_refresh (we could also
consider moving this into the ABsa/game hierarchy as the `if ext == ba2`
check is a bit off).