Hacky fix for fids created with from_tuple:
sibir: The problem appears to be that MasterMap takes a FormId or int, but it
can return a __FormId using from_tuple. The problem occurs when MasterMap is
then called on that __FormId & tries to use mod_dex on it, which isn't defined
for for __FormIds created with from_tuple.

Traceback (most recent call last):
  File "bash\gui\events.py", line 181, in _post
  File "bash\basher\dialogs.py", line 136, in DoImport
  File "bash\bosh\faces.py", line 205, in save_setFace
  File "bash\bosh\faces.py", line 326, in save_setPlayerFace
  File "bash\bosh\faces.py", line 278, in buffPackRef
  File "bash\mod_files.py", line 62, in __call__
  File "bash\brec\utils_constants.py", line 70, in mod_dex
bash.exception.StateError: mod_dex undefined for (Oblivion.esm, 022843) built from tuple

Proper fix seems to refactor so we use proper FormId instances here - we are
out of formId context when we create those and this is complex to solve
so instead I keep in the MasterMap the original masters - this will do for a
while (MasterMap is a big time producer of FormId.from_tuple, start by
refactoring faces.py uses).