Refactor irefresh to align with refresh:
I reworked the parameters of irefresh wrappers and introduced another one
(refresh_i) to track where refresh_info (sic) is passed - looks like
only refresh_i sets this (to a list or RefrData -> RefrIn instance)
and probably ShowPanel. The handling of refresh_info is encapsulated in
_list_store_dir (BAIN version) as is done for _AFileInfos. This is
WIP to allow me to absorb delete_refresh - see next commit. Note that
what=C is hacked in - we should have a proper data store for converters
as we have for bsa infos and decouple BAIN from it - laterz. Note also
booting is now kwarg only.