[!]WIP centralize/optimize heuristics:
Now the dust settled - why use filename heuristics to see which are
the plugin loaded bsas since we just calculated them? This leaves the
heuristics only for ini-loaded bsas - as I don't think
main/patch/interface logic applies to plugin-loaded bsas for other
plugins or that those bsas are loaded by the master esm as pl_bsas [!]
- if this proves the case we should repeat the ordering inside
isMissingStrings just for the master esm.