SF: Updates for v1.11+ saves
The v1.11 update appears to have added some number of unknown bytes, 14
for Arthmoor & myself. However, I've seen a report with a different
number, so, if this block is variable, this is not the right solution.
This commit cannot hit dev until this is confirmed, which will require
some decoding.

Utunmo: this will will be reworked starting at "WIP SF saves:" commit.