WIP fixups for deleted files handling and absorb applyTweakFile to base:
The slight repetition in _fmt_setting could be absorbed in saveSettings
if we drop the OBSE override ;) The former could be added parameters to
format the section and deleted settings too but that would be baroque.
Note that if the wizard knew what kind of ini we are editing
generateTweakLines could go - that'd be a wonder. See
169d8347c1e4f3a3f6d696d4a660f89987f6bffc for the origins of the code -
`if not realSection` does not make much sense but probably values is
never empty either. Anyway we would need to add fnEditObseIni and have
the user pass the ini type - what if we have an ini with a [Set]
section? We can't use get_ini_type as we need the ini text for that,
which generateTweakLines generates.
Note that this is the last use of DefaultLowerDict and no setdefault
remain.