Make List Dependent a 'highlight' dialog
Previously called these 'change highlight' dialogs, but that already
doesn't really fit. 'Item highlight' is a better term, so that's what
I'm going with now. Some refactoring/renaming to enable that.

Gets rid of a bunch of ugly code in there, an open ##: pointing out how
hacky the implementation is and is a negative LOC commit on top of all
that :)