Write out the Starfield CCC file:
See the discussion in #681. Note that we need to do much more to have
proper Starfield support - see the detailed overview at:

https://blog.ortham.net/posts/2024-06-28-load-order-in-starfield/#load-order

Note:

- Starfield.ccc may be loaded from My Games
- *plugins* may be loaded from My Games -> tough, goddbye store_dir,
this needs #336 being finalized (lots of machinery moved to the Info
classes)
- sTestFile1 through sTestFile10 (also multiple dirs the inis can hide)
- right out
- changes in plugin txt

also:

- No spec for error handling - for instance if we encounter duplicates
is the first or last occurrence that counts?

I went ahead and renamed a few variables, which made me notice that
SSE could use the Skyrim list. Could be a bit more efficient (we stat the
ccc file twice for SF for instance but would do for now)

Closes #682