Rework handling of max_esls/espms limits: RRR
These should be static apart from the fact that SkyrimVR can have esls
via an fose plugin - and to check that we must go post_init - see
ca361bd2d11ce73a481ca9998ca1da0ba68676bc and comments there. We should
probably add the dirs init in the game (as it depends on it) and absorb
post_init (always tricky/hacky to post init) but that's from the top of
my head. Meanwhile I figured out this way, that removes the pointless
wrapping in load_order and adds a couple instance attributes to game
instead. We already had the gamePath so we just have one hack (Game is
supposed to be as static as possible) rather than two (importing bush
in game and having what is essentially game info into LoGame).

Under # <--- RRR 578 and # <--- RRR 681 (hopefully this is more extendable for "medium"
plugins)