PoC PluginFlag(Enum): TTT
TTT -                ]),
+                next(it), HLayout(spacing=4, items=[*it] or [None]),

That giant docstring was due to not having the flags centralized -
note we should be using extension_forces_flags more. Still WIP.
Trickiest part to get right was the set_mod_flags overrides - will be
made final later. For the _SFPluginFlag.set_mod_flag override note: TTT

> A plugin is referred to as an update plugin if it does not have the
> light or medium flags set, has the update flag (0x200) set and has at
> least one master. Update plugins are referred to as “Update Master file”
> by the Creation Kit, whether or not the master flag is set, and it
> doesn’t matter if the plugin’s file extension is .esp, .esl or .esm.

From:
https://blog.ortham.net/posts/2024-06-28-load-order-in-starfield/#update-plugins

My understanding is that the esl flag won't be added at runtime - not sure
about the ESM flag for an esl extension but that should not matter TTT