[SF]Stop skipping overlay plugins from LO: RRR # 681
From https://blog.ortham.net/posts/2024-06-28-load-order-in-starfield/#update-plugins

"""
Starfield’s initial release loaded active update plugins without them
taking up a load order index, but as of 1.12.30.0 that was no longer
the case, so update plugins are treated as normal (a.k.a. “full”)
plugins as far as load order is concerned.
"""