Swap ModInfos._mergeable_by_type for ModInfo.merge_types: RRR
Note:
- I allow for any combination of types - this seems to be the case already
although as of now MERGE and ESL/OVERLAY do not appear in any game
together - but see # <--- RRR 596
- fileInfo is already in scope most of the time (while modInfos always:P)
- we only use the cache in checkMods - rest are lookups and are now a
dict lookup down. This was a hint that the cached checks belong to the
info in this case - this parallels the work in # <--- RRR 336 and # <--- RRR 353
- I had to pass the game_handle to isPBashMergeable in place of the
unused _mod_header_reader - I moved the static functions from the
latter to ModInfo - they belong there plus this simplified the
lambdas in error_msgs.

Mopy/bash/basher/__init__.py: the  `+ ' '` had to go